+++ carlpion/src/main/java/com/kh/carlpion/comment/model/dto/CommentReportDTO.java
@@ -20,5 +20,4 @@ public class CommentReportDTO {
 	private String content;
 	private String nickName;
 	private Date createDate;
-	private String isActive;
 }
