+++ carlpion/src/main/java/com/kh/carlpion/comment/model/dto/CommentReviewDTO.java
@@ -20,5 +20,4 @@ public class CommentReviewDTO {
 	private String content;
 	private String nickName;
 	private Date createDate;
-	private String isActive;
 }
