+++ carlpion/src/main/java/com/kh/carlpion/comment/model/service/CommentNoticeService.java
@@ -3,7 +3,6 @@ package com.kh.carlpion.comment.model.service;
 import java.util.List;
 
 import com.kh.carlpion.comment.model.dto.CommentNoticeDTO;
-import com.kh.carlpion.comment.model.vo.CommentNoticeVO;
 
 public interface CommentNoticeService {
