+++ carlpion/src/main/java/com/kh/carlpion/comment/model/service/CommentReportService.java
@@ -3,7 +3,6 @@ package com.kh.carlpion.comment.model.service;
 import java.util.List;
 
 import com.kh.carlpion.comment.model.dto.CommentReportDTO;
-import com.kh.carlpion.comment.model.vo.CommentReportVO;
 
 public interface CommentReportService {
