+++ carlpion/src/main/java/com/kh/carlpion/comment/model/service/CommentReviewService.java
@@ -3,7 +3,6 @@ package com.kh.carlpion.comment.model.service;
 import java.util.List;
 
 import com.kh.carlpion.comment.model.dto.CommentReviewDTO;
-import com.kh.carlpion.comment.model.vo.CommentReviewVO;
 
 public interface CommentReviewService {
