+++ carlpion/src/main/java/com/kh/carlpion/comment/model/vo/CommentReportVO.java
@@ -1,7 +1,5 @@
 package com.kh.carlpion.comment.model.vo;
 
-import java.sql.Date;
-
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Value;
@@ -10,11 +8,7 @@ import lombok.Value;
 @Builder
 @Getter
 public class CommentReportVO {
-	private Long commentNo;
 	private Long reportNo;
 	private Long userNo;
 	private String content;
-	private String nickName;
-	private Date createDate;
-	private String isActive;
 }
