+++ carlpion/src/main/java/com/kh/carlpion/point/model/dao/PointMapper.java
@@ -4,7 +4,9 @@ import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 
+import com.kh.carlpion.point.model.dto.LikeDTO;
 import com.kh.carlpion.point.model.dto.PointHistoryDTO;
+import com.kh.carlpion.point.model.vo.LikeVO;
 import com.kh.carlpion.point.model.vo.PointHistoryVO;
 import com.kh.carlpion.point.model.vo.PointVO;
 
@@ -18,4 +20,10 @@ public interface PointMapper {
 	void updateUserPoint(PointVO pointVO);
 	
 	void updateUserLevel(PointVO pointVO);
+	
+	void saveReviewLike(LikeVO likeVO);
+	
+	List<LikeDTO> findAll();
+	
+	void deleteReviewLike(LikeVO likeVO);
 }
