+++ carlpion/src/main/java/com/kh/carlpion/point/model/service/PointService.java
@@ -2,6 +2,7 @@ package com.kh.carlpion.point.model.service;
 
 import java.util.List;
 
+import com.kh.carlpion.point.model.dto.LikeDTO;
 import com.kh.carlpion.point.model.dto.PointDTO;
 import com.kh.carlpion.point.model.dto.PointHistoryDTO;
 
@@ -14,4 +15,10 @@ public interface PointService {
 	void updateUserPoint(PointDTO pointDTO);
 			
 	void updateUserLevel(PointDTO pointDTO);
+	
+	void saveReviewLike(LikeDTO likeDTO);
+	
+	List<LikeDTO> findAll();
+	
+	void deleteReviewLike(LikeDTO likeDTO);
 }
