+++ carlpion/src/main/java/com/kh/carlpion/point/model/vo/PointHistoryVO.java
@@ -1,7 +1,5 @@
 package com.kh.carlpion.point.model.vo;
 
-import java.sql.Date;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,10 +10,8 @@ import lombok.Value;
 @Getter
 @AllArgsConstructor
 public class PointHistoryVO {
-	private Long historyNo;
 	private Long userNo;
 	private Long reviewNo;
 	private Long pointChange;
 	private String reason;
-	private Date createDate;
 }
