+++ carlpion/src/main/java/com/kh/carlpion/report/model/dto/ReportDTO.java
@@ -19,8 +19,9 @@ public class ReportDTO {
 	private String title;
 	private String content;
 	private String nickName;
-	private String fileUrl;
 	private Date createDate;
 	private Long count;
-	private String isActive;	
+	private String fileUrl;
+	private Long point;
+	private String userLevel;
 }
