+++ carlpion/src/main/java/com/kh/carlpion/review/model/service/ReviewServiceImpl.java
@@ -11,6 +11,8 @@ import com.kh.carlpion.auth.model.service.AuthService;
 import com.kh.carlpion.exception.exceptions.NotFindException;
 import com.kh.carlpion.exception.exceptions.UnauthorizedException;
 import com.kh.carlpion.file.service.FileService;
+import com.kh.carlpion.point.model.dto.PointHistoryDTO;
+import com.kh.carlpion.point.model.service.PointService;
 import com.kh.carlpion.review.model.dao.ReviewMapper;
 import com.kh.carlpion.review.model.dto.ReviewDTO;
 import com.kh.carlpion.review.model.vo.ReviewVO;
@@ -26,6 +28,7 @@ public class ReviewServiceImpl implements ReviewService {
 	private final ReviewMapper reviewMapper;
 	private final AuthService authService;
 	private final FileService fileService;
+	private final PointService pointService;
 
 	@Override
 	@Transactional
@@ -43,6 +46,7 @@ public class ReviewServiceImpl implements ReviewService {
 									   .content(reviewDTO.getContent())
 									   .build();
 		reviewMapper.save(requestData);
+		Long reviewNo = requestData.getReviewNo();
 		
 		if(files != null && !files.isEmpty()) {
 			for(MultipartFile file : files) {
@@ -51,14 +55,14 @@ public class ReviewServiceImpl implements ReviewService {
 					String filePath = fileService.storage(file);
 					
 					ReviewVO requestFileData = ReviewVO.builder()
-													   .reviewNo(requestData.getReviewNo())
+													   .reviewNo(reviewNo)
 													   .fileUrl(filePath)
 													   .build();
 					reviewMapper.saveFile(requestFileData);
 				}
 			}
 		}
-//		log.info("save: {}", requestData);
+//		log.info("save: {}", requestData);		
 	}
 
 	@Override
