+++ carlpion/src/main/java/com/kh/carlpion/review/model/vo/ReviewVO.java
@@ -1,7 +1,5 @@
 package com.kh.carlpion.review.model.vo;
 
-import java.sql.Date;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -18,7 +16,4 @@ public class ReviewVO {
 	private String content;
 	private String nickName;
 	private String fileUrl;
-	private Date createDate;
-	private Long count;
-	private String isActive;
 }
