+++ carlpion/src/main/resources/mapper/CommentNoticeMapper.xml
@@ -2,21 +2,21 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.kh.carlpion.comment.model.dao.CommentReportMapper">
+<mapper namespace="com.kh.carlpion.comment.model.dao.CommentNoticeMapper">
 	<insert 
 		id="saveComment"
-		parameterType="com.kh.carlpion.comment.model.vo.CommentReportVO"
+		parameterType="com.kh.carlpion.comment.model.vo.CommentNoticeVO"
 	>
 		INSERT 
-		  INTO	CP_RP_COMMENT
+		  INTO	CP_NT_COMMENT
 		  	 (
-				REPORT_NO
+				NOTICE_NO
 			 ,	USER_NO
 			 , 	CONTENT
 			 )
 		VALUES 
 			 (
-			   	#{reportNo}
+			   	#{noticeNo}
 			 , 	#{userNo} 	
 			 , 	#{content} 	
 			 )
@@ -24,17 +24,16 @@
 	
 	<select 
 		id="findAllComment" 
-		resultType="com.kh.carlpion.comment.model.dto.CommentReportDTO"
+		resultType="com.kh.carlpion.comment.model.dto.CommentNoticeDTO"
 	>
 		SELECT 
 				c.COMMENT_NO	AS	commentNo
-			 ,	c.REPORT_NO		AS	reportNo
-			 ,	c.USER_NO		AS	userNo
+			 ,	c.NOTICE_NO		AS	noticeNo
 			 ,	c.CONTENT
 			 ,	u.NICKNAME		AS	nickName
 			 ,	c.CREATE_DATE	AS	createDate
 		  FROM
-		  		CP_RP_COMMENT c
+		  		CP_NT_COMMENT c
 		  JOIN
 		  		CP_USER u 	ON 	u.USER_NO = c.USER_NO
 		 WHERE	
@@ -42,7 +41,7 @@
 		 ORDER
 		 	BY	
 		 		c.CREATE_DATE DESC
-	</select>
+	</select>	
 	
 	<select 
 		id="findByUserNo"
@@ -52,13 +51,13 @@
 		SELECT
 				USER_NO		AS	userNo
 		  FROM	
-		  		CP_RP_COMMENT
+		  		CP_NT_COMMENT
 		 WHERE
 		 		COMMENT_NO = #{commentNo}	  		
 	</select>
 	<update id="softDeleteCommentById">
 		UPDATE
-				CP_RP_COMMENT
+				CP_NT_COMMENT
 		   SET
 		   		IS_ACTIVE = 'N'
 		 WHERE
