+++ carlpion/src/main/resources/mapper/CommentReviewMapper.xml
@@ -29,7 +29,6 @@
 		SELECT 
 				c.COMMENT_NO	AS	commentNo
 			 ,	c.REVIEW_NO		AS	reviewNo
-			 ,	c.USER_NO		AS	userNo
 			 ,	c.CONTENT
 			 ,	u.NICKNAME		AS	nickName
 			 ,	c.CREATE_DATE	AS	createDate
