+++ carlpion/src/main/resources/mapper/NoticeMapper.xml
@@ -49,7 +49,6 @@
 	>
 		SELECT 
 				n.NOTICE_NO		AS	noticeNo
-			 ,	n.USER_NO		AS	userNo
 			 ,	n.TITLE
 			 ,	n.CONTENT
 			 ,	u.NICKNAME		AS	nickName
@@ -58,9 +57,9 @@
 		  FROM
 		  		CP_NOTICE n
 		  JOIN
-		  		CP_USER u 	ON 	u.USER_NO = n.USER_NO
+		  		CP_USER u 		ON 	u.USER_NO = n.USER_NO
 		  JOIN
-		  		CP_POINT p 	ON 	p.USER_NO = n.USER_NO	
+		  		CP_POINT p 		ON 	p.USER_NO = n.USER_NO	
 		 WHERE	
 		 		n.IS_ACTIVE = 'Y'
 		 ORDER
@@ -85,20 +84,22 @@
 	>
 		SELECT 
 				n.NOTICE_NO		AS	noticeNo
-			 ,	n.USER_NO		AS	userNo
 			 ,	n.TITLE
 			 ,	n.CONTENT
 			 ,	u.NICKNAME		AS	nickName
 			 ,	n.CREATE_DATE	AS	createDate
 			 ,	n.COUNT
+			 ,	f.FILE_URL		AS	fileUrl
 			 ,	p.POINT
-			 ,	p.USER_LEVEL
+			 ,	p.USER_LEVEL	AS	userLevel	
 		  FROM
 		  		CP_NOTICE n
 		  JOIN
-		  		CP_USER u 	ON 	u.USER_NO = n.USER_NO
+		  		CP_USER u 		ON 	u.USER_NO = n.USER_NO
 		  JOIN
-		  		CP_POINT p 	ON 	p.USER_NO = n.USER_NO
+		  		CP_POINT p 		ON 	p.USER_NO = n.USER_NO
+		  JOIN
+		  		CP_NT_FILE f 	ON 	f.NOTICE_NO = r.NOTICE_NO
 		 WHERE	
 		 		n.IS_ACTIVE = 'Y'
 		   AND	
@@ -142,7 +143,7 @@
 		id="findFileByAll"
 		resultType="string"
 	>
-		SLELCT
+		SELECT
 				FILE_URL	AS	fileUrl
 		  FROM
 		  		CP_NT_FILE
