+++ carlpion/src/main/resources/mapper/PointMapper.xml
@@ -31,7 +31,6 @@
 	>
 		SELECT
 				HISTORY_NO
-			 ,	USER_NO
 			 ,	REVIEW_NO
 			 ,	POINT_CHANGE
 			 ,	REASON
@@ -64,4 +63,43 @@
 		 WHERE
 		 		USER_NO = #{userNo}
 	</update>
+	
+	
+	<insert 
+		id="saveReviewLike"
+		parameterType="com.kh.carlpion.point.model.vo.LikeVO"
+	>
+		INSERT 
+		  INTO	
+		  		CP_LIKE
+		VALUES 
+			 (
+			   	#{userNo}
+			 , 	#{reviewNo} 		
+			 )
+	</insert>
+	
+	<select 
+		id="findAll" 
+		resultType="com.kh.carlpion.point.model.dto.LikeDTO"
+	>
+		SELECT 
+			 	USER_NO		AS	userNo
+			 ,	REVIEW_NO	AS	reviewNo
+		  FROM
+		  		CP_LIKE
+	</select>
+
+	<delete 
+		id="deleteReviewLike"
+		parameterType="com.kh.carlpion.point.model.vo.LikeVO"
+	>
+		DELETE
+		  FROM
+		  		CP_LIKE
+		 WHERE	
+		 		USER_NO = #{userNo}
+		   AND
+		 		REVIEW_NO = #{reviewNo}
+	</delete>	
 </mapper>
