--- carlpion/src/main/resources/pointMapper/LikeMapper.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.kh.carlpion.point.model.dao.LikeMapper">
-	<insert 
-		id="saveReviewLike"
-		parameterType="com.kh.carlpion.point.model.vo.LikeVO"
-	>
-		INSERT 
-		  INTO	
-		  		CP_LIKE
-		VALUES 
-			 (
-			   	#{userNo}
-			 , 	#{reviewNo} 		
-			 )
-	</insert>
-	
-	<select 
-		id="findAll" 
-		resultType="com.kh.carlpion.point.model.dto.LikeDTO"
-	>
-		SELECT 
-			 	USER_NO		AS	userNo
-			 ,	REVIEW_NO	AS	reviewNo
-		  FROM
-		  		CP_LIKE
-	</select>
-
-	<delete 
-		id="deleteReviewLike"
-		parameterType="com.kh.carlpion.point.model.vo.LikeVO"
-	>
-		DELETE
-		  FROM
-		  		CP_LIKE
-		 WHERE	
-		 		USER_NO = #{userNo}
-		   AND
-		 		REVIEW_NO = #{reviewNo}
-	</delete>
-</mapper>
